Update GeoQuery.ts
I implemented this change to address the following problem:

Prerequisites:
Test with no network connectivity

Steps:
1.) Start a location update for a specific location
2.) Follow up with an update criteria call. (This has to happen before step 1, hence the no network connectivity scenario being simulated)

Note that, in reality, there are ~30 geofire keys are to be returned when there is network connectivity.

Actual:
A 'ready' event is falsely fired for the second call, even though the db was not queried yet. The code is ignoring the current queries ongoing for the second call, and thinks that there are no outstanding queries when the second call is invoked because of this line that I replaced:
`this._outstandingGeohashReadyEvents = geohashesToQuery.slice();` 

Expected: (After this change)
The 'ready' event is only fired when all existing queries have been returned.